Glue posts handling
The export returns a map iterator with exportify_post()
which adds some useful post fields.

Then the XML generator takes it and returns another
map generator, which generates XML out of the post
object.

Now when the writer loops over the posts it has no
idea what happens behind the scenes. It looks as if
it got just an array of strings with XML in them.